refactor: [#153] UnitInfo 파일 분리, 줄간격 조정
